third_party/blink/web_tests/platform/generic/external/wpt/fetch/api/request/request-reset-attributes.https-expected.txt

This is a testharness.js-based test.
FAIL Request.isReloadNavigation is reset with non-empty RequestInit assert_equals: expected "old: false, new: false" but got "old: undefined, new: undefined"
PASS Request.isHistoryNavigation is reset with non-empty RequestInit
PASS Request.mode is reset with non-empty RequestInit when it's "navigate"
Harness: the test ran to completion.
